pyproject.toml
[project]
name = "occam-backend"
version = "0.1.0"
description = "Occam Backend - Knowledge Management System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lark-oapi>=1.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "instructor>=0.4.0",
    "pydantic>=2.0.0",
    "notion-client>=2.0.0",
    "httpx>=0.25.0",
    "markdownify>=0.11.0",
]
[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true